Add JMX minibankc-xmanamgement for managing application
