+++ src/main/java/com/example/minibankc/MinibankcApplication.java
@@ -1,55 +1,20 @@
 package com.example.minibankc;
 
-import com.github.benmanes.caffeine.cache.Cache;
-import com.github.benmanes.caffeine.cache.Caffeine;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.i18n.LocaleContextHolder;
 import org.springframework.context.support.ResourceBundleMessageSource;
 import org.springframework.web.servlet.LocaleResolver;
-import org.springframework.web.servlet.i18n.LocaleChangeInterceptor;
 import org.springframework.web.servlet.i18n.SessionLocaleResolver;
 
 import java.util.Locale;
-import java.util.concurrent.TimeUnit;
 
 @SpringBootApplication
 public class MinibankcApplication {
 
-	public static void main(String[] args) {
+    public static void main(String[] args) {
 
-		SpringApplication.run(MinibankcApplication.class, args);
-	}
-
-	/**
-	 * add internationalization to a Spring Boot application
-	 * In order for our application to be able to determine which locale is currently being used,
-	 * we add a LocaleResolver bean
-	 * @return localeResolver
-	 */
-	@Bean
-	public LocaleResolver localeResolver() {
-		SessionLocaleResolver localeResolver = new SessionLocaleResolver();
-		localeResolver.setDefaultLocale(Locale.US);
-		return localeResolver;
-
-	}
-
-	/**
-	 * Defining the Message Sources
-	 * By default, a Spring Boot application will look for message files containing
-	 * internationalization keys and values in the src/main/resources folder.
-	 * @return
-	 */
-	@Bean
-	public ResourceBundleMessageSource messageSource() {
-		ResourceBundleMessageSource messageSource = new ResourceBundleMessageSource();
-		messageSource.setUseCodeAsDefaultMessage(true);
-		messageSource.setBasenames("messages");
-		messageSource.setDefaultEncoding("UTF-8");
-		return messageSource;
-	}
+        SpringApplication.run(MinibankcApplication.class, args);
+    }
 
 }
