--- src/main/java/com/example/minibankc/config/InternationalConfig.java
@@ -1,34 +1,0 @@
-package com.example.minibankc.config;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import org.springframework.web.servlet.i18n.LocaleChangeInterceptor;
-
-/**
- * @author Mahdi Sharifi
- * @version 2022.1.1
- * https://www.linkedin.com/in/mahdisharifi/
- * @since 4/30/22
- */
-@Configuration
-@Slf4j
-public class InternationalConfig implements WebMvcConfigurer {
-
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(localeChangeInterceptor());
-        log.info("#addInterceptor: " + registry);
-    }
-
-    @Bean
-    public LocaleChangeInterceptor localeChangeInterceptor() {
-        LocaleChangeInterceptor lci = new LocaleChangeInterceptor();
-        lci.setParamName("lang");
-        log.info("#language change to: " + lci.getParamName());
-        return lci;
-    }
-}
