+++ src/main/java/com/example/minibankc/config/InternationalizationConfig.java
@@ -1,0 +1,49 @@
+package com.example.minibankc.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.support.ResourceBundleMessageSource;
+import org.springframework.web.servlet.LocaleResolver;
+import org.springframework.web.servlet.i18n.SessionLocaleResolver;
+
+import java.util.Locale;
+
+/**
+ * @author Mahdi Sharifi
+ * @version 2022.1.1
+ * https://www.linkedin.com/in/mahdisharifi/
+ * @since 5/1/22
+ */
+@Configuration
+public class InternationalizationConfig {
+    /**
+     * add internationalization to a Spring Boot application
+     * In order for our application to be able to determine which locale is currently being used,
+     * we add a LocaleResolver bean
+     *
+     * @return localeResolver
+     */
+    @Bean
+    public LocaleResolver localeResolver() {
+        SessionLocaleResolver localeResolver = new SessionLocaleResolver();
+        localeResolver.setDefaultLocale(Locale.US);
+        return localeResolver;
+
+    }
+
+    /**
+     * Defining the Message Sources
+     * By default, a Spring Boot application will look for message files containing
+     * internationalization keys and values in the src/main/resources folder.
+     *
+     * @return
+     */
+    @Bean
+    public ResourceBundleMessageSource messageSource() {
+        ResourceBundleMessageSource messageSource = new ResourceBundleMessageSource();
+        messageSource.setUseCodeAsDefaultMessage(true);
+        messageSource.setBasenames("messages");
+        messageSource.setDefaultEncoding("UTF-8");
+        return messageSource;
+    }
+}
