+++ src/main/java/com/example/minibankc/util/apects/RateControllersAspectHandler.java
@@ -1,4 +1,4 @@
-package com.example.minibankc.util.filter;
+package com.example.minibankc.util.apects;
 
 import com.example.minibankc.exception.BadRequestAlertException;
 import com.github.benmanes.caffeine.cache.Cache;
