--- src/main/resources/templates/international.html
@@ -1,21 +1,0 @@
-<html xmlns:th="https://thymeleaf.org">
-<h1 th:text="#{greeting}"></h1>
-<span th:text="#{lang.change}"></span>:
-<select id="locales">
-    <option value=""></option>
-    <option value="en" th:text="#{lang.eng}"></option>
-    <option value="nl" th:text="#{lang.nl}"></option>
-</select>
-
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.1.1/jquery.min.js">
-</script>
-<script type="text/javascript">
-    $(document).ready(function() {
-        $("#locales").change(function () {
-            var selectedOption = $('#locales').val();
-            if (selectedOption != ''){
-                window.location.replace('international?lang=' + selectedOption);
-            }
-        });
-    });
-</script>
